Added function to open URL and a base Android activity.
To open an internet URL, use openURL() function on Device class. URL can be opened on a browser or on a WebView based on the parameter used. Right now opening URL on a Webview doesn't do anything though.

Most of Android initialization is moved to GameActivity class so developers just need to extend it and initialize a few things. Check out the example activity included to see the initialization needed.