Added opening URL on WebView instead of browser.
openURL() now works on WebView, developers can use WebActivity class provided in the framework or create their own. The WebActivity must be registered on the manifest and the gameplay activity needs to launch as SingleTop to allow player to open the URL in a webview.

openURL() function can be accessed via Utilities class with parameters to set the title and loading label.