Separated sample project and compiled as JAR. Version 0.9!
The example and the framework is now two separate projects to make it clearer which part is the framework and which one isn't. Android parts of the project (like icons and resources) are now removed from the framework as well.

By moving the example codes to a separate project, the framework can now be compiled as a JAR file. This allows the framework to be used simply by adding the library (the JAR file) to the game project build path instead of setting up the library as an Android project and referencing to it. The old

And also, current version of the framework is now officially labeled as version 0.9!