Changed how base width and height work.
Base width and height can now be accessed from Utilities class. Just like the previous version, the framework will choose the side with positive number as reference, but in this version, the framework would convert those negative number into positive. So, if the reference size is 320 x 480 and width is going to be used as the reference, set baseWidth as 320 and baseHeight as -480, so the framework would return the correct base height (480).

Also included the latest compiled library (let's call it developer version) in the release folder , the one without version number. The developer version would contain incremental updates that may work differently with the official, stable one. The developer version is not officially supported, the example project would only use the stable version of the library, so it is highly recommended to use the stable version instead of the developer version.