Added framework for J2ME
As per my friend request, I added the totally incompatible J2ME
framework. This stuff is pretty much unusable (or at least require a lot
of tinkering), but it's a good place to start for anyone looking to
develop games in J2ME.

Unlike its Android counterpart, this isn't a separate library, but more
like a base project which could be extended with various game logics. I
intend to make it into a separate library though, IF I have the time.