Separated Device class from GameCanvas class on J2ME framework.
Now the Device class of the J2ME framework doesn't extend the GameCanvas anymore. Instead, the GameCanvas is now extended on ArkCanvas, and to access it, developers can use the getCanvas() function from the Midlet. The Device class itself can still be used as usual.

Also added a base MIDlet class so developers don't need to go through all of the application setup. Check out the sample project on how to use the base MIDlet class.

Now the application can also exit properly when there's no more active state in the stack.