Added the capability to crop Label [Android]
Now Label can be cropped so it ony draws a part of itself instead of the
whole label. To crop the label, use the function setRegion() on the
Label class to set which part of the label that will be drawn.

This is the first step towards version 0.9.1 featuring scrollable
components, since the scrollable content must be able to be cropped if
it doesn't fit the scroll window.