Fixed an issue with screen lock and rotation.
When players have lockscreen with different orientation from the game orientation, locking the device while playing and unlocking it again will call the onSurfaceChanged() function and made the game register the wrong screen size. With this commit, the onSurfaceChange() function now would check first if the activity is visible before saving the new screen size.