Fixed a cropping bug and added font configuration to Button.
Corrected some cropping values calculation so the size of the cropping cannot be bigger than the original size of the component.

Also added some font parameters to the Button class constructor so the font for the button text can be changed from the default font. The font setter on Button class is removed so the font can only be configured when the button is instantiated because changing font after the button is created is cumbersome.