Made items in ItemList more publicly accessible.
Now ItemList class have accessor functions that give access the number of items on that list and the item on a certain position. UpdateItems function is also made public so any changes to the individual item can be propagated to the list manually without waiting for the next update or scrolling event.