Changed some components: List, Scrolling Image, and Sprite.
Setting ItemList position will now also properly change the position of items as well. And following Image class, now Sprite class extends Croppable class as well, making it possible to be used as one of the item for ItemList. ScrollingImage used to have a 1 pixel gap on some resolutions due to a rounding error. To fix it, a -1 offset has been added when drawing ScrollingImage interface component.