Added Image and Sprite rotation with custom pivot.
Now both Image and Sprite class can be rotated around a pivot point other than the center. To add a pivot point to a rotation, simply add the X and Y coordinate of the pivot to the setRotation() function. Pivot point can be put outside of the image area too.