Added capability to the ResourceManager to load json and textures from device storage.
Now the ResourceManager class can load texture and JSON files from place other than the game assets, which is from the external or internal storage. To queue the resource, simply put the full file path as the name and set the external parameter to be true on the resource adding functions. The file can be loaded from both internal or external storage as long as the file path is correct.